[Feat/47]  맞춤 정보 API 수정

+++ data/src/main/java/com/nexters/knownknowns/data/remote/service/UserService.kt
@@ -8,7 +8,7 @@ import retrofit2.http.Path
 interface UserService {
     @PUT("api/users/{userId}")
     suspend fun putUserInfo(
-        @Path(value = "userId") userId: Int,
+        @Path(value = "userId") userId: Long,
         @Body body: UserInfoRequest,
     )
 }
